chore(mobile): apply upstream lint rules from batches 15/20 to fork Dart code
Upstream immich-30370 (formatting) and immich-30373 (Flutter + known issues) enable ~25 new
Dart lint rules. `dart analyze --fatal-infos lib test` is a CI gate, so fork-only
and fork-modified Dart had to follow.

Mostly `dart fix --apply` (directives_ordering, noop_primitive_operations,
unnecessary_breaks, unnecessary_null_checks, prefer_const_declarations,
always_declare_return_types, avoid_void_async, use_decorated_box,
combinators_ordering), plus one hand-fixed use_named_constants.

`always_put_control_body_on_new_line` stays disabled in mobile/analysis_options.yaml
per the fork's existing decision (7060ec3abdb), so its ~97 sites are untouched.

Batch 24's correctness rules (discarded_futures, cast_nullable_to_non_nullable,
close_sinks, cancel_subscriptions) are NOT enabled yet — that sweep belongs to the
next arc.

+++ mobile/test/presentation/pages/drift_remote_album_page_test.dart
@@ -6,11 +6,11 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:immich_mobile/domain/models/album/album.model.dart';
 import 'package:immich_mobile/domain/models/asset/base_asset.model.dart';
+import 'package:immich_mobile/domain/models/settings_key.dart';
 import 'package:immich_mobile/domain/models/store.model.dart';
 import 'package:immich_mobile/domain/models/timeline.model.dart';
 import 'package:immich_mobile/domain/models/timeline_temporal_scope.model.dart';
 import 'package:immich_mobile/domain/models/user.model.dart';
-import 'package:immich_mobile/domain/models/settings_key.dart';
 import 'package:immich_mobile/domain/services/remote_album.service.dart';
 import 'package:immich_mobile/domain/services/store.service.dart';
 import 'package:immich_mobile/domain/services/timeline.service.dart';
@@ -70,7 +70,7 @@ TimelineService _service(List<Bucket> buckets) {
   return TimelineService((
     bucketSource: () => Stream.value(buckets),
     assetSource: (offset, count) async {
-      final end = (offset + count).clamp(0, assets.length).toInt();
+      final end = (offset + count).clamp(0, assets.length);
       if (offset >= end) return const <BaseAsset>[];
       return assets.sublist(offset, end);
     },
